Implemented bare-bones runtime VIs.
This commit introduces the Experimental Configuration control type
definition, which defines the data that will pass through the ex-
perimental queue, as well as introducing the dequeue-ing VI and the
VI that executes the experiments as described by dequeued config.

The Raw Data Interpretor VI is only implemented for its interface,
discussion is necessary to determine in what manner the raw sensor
data should be calibrated and transformed.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>