Implemented sempahore for synchronizing config PUTs
In order to ensure that the system is user-friendly
and easy to reason about, it is necessary to ensure
that experiments interact with the API in an atomic
manner. This commit introduces a semaphore that will
be used to synchronize any actions that:
  - enqueue experiments to be executed
  - change configurations that impact experiments

This commit also refactors existing configuration
menus to use the semaphore.

Fixes issue #39

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>