Built set-point enforcer VIs
This commit introduces the first implementation of
the basic concept that will be used to implement the
entire experimentation flow. The Set-Point Driver VI
continually attempts to acquire the enqueue semaphore
in order to push to the experimental queue a short
`experiment' that forces the set-points and records
current sensor state. Message-passing is used for
control of the loop from the main VI.

The initial interface for the experiment VIs has no
start or finish hooks, as it seems LabVIEW does not
call SubVIs asynchronously by default, so they are
not necessary. Instead the VIs take the queue as the
only input.

Fixes #40

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>