Reconfigured asynchronous calls to be `call-and-collect'
While the behavior we would *like* here is `call-and-
forget', possible with the 0x80 option code, that mode
of asynchronous calling has the side effect that the
called VI does not stop when the calling VI is aborted
or finishes.

This means that in order to stop all of the asynchronous
VIs, it is either necessary to open them all by hand and
abort them or to close the entire project. This is less
than ideal.

Therefore, the 0x100 `call-and-collect' code is used
instead. This is not the intended use of this code and
LabVIEW documentation suggests against this as the
LabVIEW runtime keeps memory allocated for the VI even
after it's finished running until the collect happens.

This is not an issue for us as all of the VIs we are
running in this VI are unparameterized, yield no output,
and run forever. All of these VIs will be running up until
the parent VI is closed, anyway, so there is no additional
memory footprint.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>