Implemented the raw data interpreter VI
To interpret the raw voltage data, we assume
that the input channels are in the following
order:
  - 0: Z piezo
  - 1: DIV signal
  - 2: SUM signal
  - 3: X piezo (if applicable)
  - 4: Y piezo (if applicable)

For the X,Y, and Z signals we can simply
multiply the voltage signal (V) by the piezo
response (nm/V) to get the position (nm).

For the DIV signal, we multiply it by both the
slope (nm/V) and spring constant (pN/nm)
to get the force (pN) on the cantilever.

We pass the SUM signal through untouched.

Fixes #43

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>