Implemented a debuging sidecar VI
In order to debug the signals being sent to the DAQ
in real time, a second queue is built and pushed to
by the experimental dequeuer VI. This queue has the
capacity for only one item, which ensures that even
if there is no debugging VI actively dequeueing any
bundles, the total memory footprint of the queue is
always only one bundle, as items are enqueued using
only the `lossy enqueue` method.

This debugging sidecar VI reads from this new queue
and pushes the output waveforms for the X, Y, and Z
channels into waveform charts.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>