Implemented a versioned config serialization API
In order to allow users to load and save configuration
for the AFM controller, a serialization API was neces-
sary. LabVIEW natively only supports serialization of
types into TDMS and XML. In order to allow users to
inspect and modify their serialized configurations on
disk without having to develop a VI or other interface
for the task, XML was chosen.

Once 1.0.0 is cut, we will honor a strict backwards
compatibility guarantee on the serialization API. It is
yet undecided how this gurantee will apply to the
internal API dictated by the {GET/PUT/DELETE} Funct-
ional Global Variable VIs. The guarantee will be hon-
ored by the serialization API by keeping the XML-to-
data and data-to-XML VIs versioned and by requiring a
version with any serialized API object. When the ver-
sion is bumped, a new directory will be made under
`src/Serialization/Configuration' and new VIs will be
placed there.

A user event is unfinished but necessary to make the
configuration loading process friendly for users. A
new user event will be created and fired after config-
uration is loaded or an experiment is finished. All UI
plugins displaying configuration information will wait
on this event and upadate their views when it occurs.

Fixes #42

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>