Added tests for experiment and set-point drivers
Fiddled with the in-place debugging visuals to make
them easier to use and more useful as well. It is
impossible to increase the size of the buffer used
to show the in-place debugging chart, so I set the
buffer size to be 30,000 points in hopes that this
will be enough. If a user decided to run experiments
with more than 10,000 points, the in-place debugger
will not be very useful as they will not see much.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>