Moved raw data interpretation into experiments
Since we allow for experiments to indicate which
channels to use for DAQ, it is necessary to also
allow each experiment to interpret the raw data
resulting from the experiment itself.

The experiment directory has been refactored to
reflect the current structure:
  - program-supported defaults are grouped together
  - experiments are bundled into subdirectories

Further refactoring to move the experimental
parameters functional globals into the experiment
directories and dynamically populate that sub-
panel on the main GUI is necessary.

Fixes #48
Fixes #49
Fixes #57
Fixes #61

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>