Implemented a `save' button for experimental data
Each experiment is now expected to provide the VI to
be used to serialize experimental data from the last
experimental run. A `save' button now is displayed
underneath the `Run Experiment' button that executes
the VI and provides it the path to the save file
location. The provided path is lacking of an exten-
sion, to allow for different save VIs to save diff-
erent types of data.

Fixes #8
Fixes #44

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>