Implemented slope calibration VIs
In order to allow for the calibration of the slope (nm/V) and
the spring constant (pN/nm) for the cantilevers, the slope
calibration VI runs a single Z ramp in approach mode to get a
plot of signal deviation (V) as a function of Z position (nm)
The inverse slope of the most linear portion of this plot as
selected by the user is recorded as the `slope' we are after.

The spring constant is calculated by multiplying the previous
power peak (Vrms^2) by the square of the slope (nm^2/V^2) and
dividing room-temperature KbT (nm*pN) by this value to yield
the spring constant in pN/nm.

Fixes #4
Fixes #26

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>