Formatted force and Z position data better for single ramp
For both the force and Z position data gathered in the
single ramp experiments, we need the data to be mirrored
as well as to have reasonable values. Since all of the
data is relative, it does not matter if the values that
are displayed are shifted. Initially, we simply invert
all of the records and zero them out. In the future, we
will want to dynamically determine the force threshold
and zero that trace out at the threshold.

Fixes #74

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>