Added output waveform coercion for output value safety
In order to ensure that we never try to emit output
waveforms that are either unsafe for the piezos that
are in use, or for the DAQ card that is driving them,
two levels of waveform coercion are added. Coercion
for piezo limits is done in waveform generation, as
this is the only place that a waveform is directly
associated with a piezo. Output voltage limit
coercion is done in the experiment executor, as this
is the only place where we can be sure to catch all
waveforms, even if they themselves did not coerce on
creation.

Fixes #41
Fixes #82

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>